refactor: eliminated and segregated code
